Use devise to sign in the test user.
Creating a new comment or editing an existing one requires users to be signed in/authenticated (as specified in the comments controller).
Without signing in, trying to request the new & edit actions from the controller would fail the authentication and return false. This is not what the test was expecting.