Updated comment views tests.
Fixed expectation messages for each test. Each still stated that it was
trying to render the 'create' view, though this is incorrect and there
is no longer a 'create' view. Also added expectations for what will be
rendered on the page for the comment 'show' action to ensure that the
comment contents are being correctly displated.