Add tag controller and helper files
Base controller & helper files with auto-generated rspec tests. The old
way of handling tags needs to be migrated over and shuffled around a
little bit. For now there are no 'admin_tags' so there won't be a need
to have a separate module for sharing things between them and regular
'tags.'