Add tag view templates
Tags only have an index template to show all tags currently stored in
the database as well as two partials. The 'tag' partial contains the
name of the tag and will eventually be a link to search for ideas with
that tag. The second partial, 'tag_editor,' contains a portion of a
form that will be included with the new idea/edit idea forms.