Methods for finding/creating Tags
There should be no duplicate tags stored in the database. Tags are
retrieved as strings after being stripped of any extra characters. If
the resulting string already exists in the Tags table, add that result.
If it does not, create a new one and add that to the tags array
instead.