Change confirmable Devise feature spec expectations
Feature tests for Devise need to change expectations of what will be
shown after signing up for an account. Additionally, a test user needs
to be confirmed in order to sign in and be redirected to the correct
page.

Test users are also changed to use the user factory for generating test
data.