Fix Tag creation
Pass only the string of incoming tags from the Idea controller to the
'from_string' Tag class method. The method itself will handle the
stripping and splitting of the string into tag names.

Change the order in which string stripping and splitting is done in the
tags in 'Tag.from_string.' The string should be stripped and downcased
first before being split into an array.

After splitting, each substring from the array is passed to the
'find_or_create_by(name: ...)' method. This method replaces the
deprecated 'find_or_create_by_name(...)' used previously. This will
find an existing tag record with the same name as the string argument
if it exists; if it doesn't a new one will be created. These records
are added to the 'tags' array to be returned to the Idea controller.