Add 'name' to strong parameters for Devise
Since strong parameters are handled in the controllers in Rails 4,
devise needs to be told to allow the 'name' field in the user
registration form to be passed to the model for record creation. If
this isn't done, the field may be filled in by the user but the
parameter won't make it to the model and record creation will fail with
an empty name.