Bring over the rest of the home#index template
Add the rest of the Rails 2 home#show template to the new application's
home#index template. Paths have been adjusted to reflect the changes in
some of the new names. The ability to map ideas based on their location
isn't added and currently isn't planned to be, so the part of the
template containing that content has been ignored.

Searching for ideas ideas based on their newness or popularity is still
not implemented.