Set '@user_vote' variable in IdeasController
Set a variable containg the record of a user's vote on an idea if the
current user has voted on the idea who's view is being rendered. Check
to see if a user is signed in first, then if they have voted on the
idea in the first place before trying to retrieve it.