Add some (breaking) Vote feature specs
Create some feature specs for testing voting functionality. Currently,
tests break for one reason or another when all four are enabled.
One issue is that these specs expect to be executed in a particulr
order, which they shouldn't be so they'll have to be adjusted.
Additionally, the votes controller appears to be behaving a little
strangely. The view rendered after voting doesn't always appear to be
up to date and needs to be refreshed.

TODO:
   * Fix Votes controller/Idea vote counting to render correct content
   * Fix Vote feature specs to perform the correct actions for testing