Specify templated for rendering in Currents specs
Specify which templates are to be rendered in the Currents view specs.
By moving all view specs into the same file, rspec tries to find a view
called 'CurrentsViews' when simply calling 'render.' In order to test
all the views in one file, the template to be rendered needs to be
passed as an argument to the 'render' method.