Add twitter environment variable loading
Create files to store and load the environment settings for twitter
authentication. Admins will set their Twitter API keys and secrets in
the .yml file and these settings will be loaded by the initializer.

Note: Initializers are loaded in alphabetical order by Rails on server
start-up. Because the twitter settings will need to be made available
in the devise configuration, the filename is prepended with '01_' to
move the twitter initializer ahead of the devise one. If the twitter
intializer was left as just 'twitter.rb,' the 'TWITTER_CONFIG' global
would not be unset/nil during devise intialization.