Finish Twitter omniauth
Solution for signing in Omniauthed users should be generic and allow
for new strategies to be added later. After user credentials have been
authenticated with the third party, information is passed back to the
controller and handled generally. User records are created or found
based on the provider/uid combination, which should be unique.