Check if provider is enabled before adding link
Determine if each provider is enabled in 'config/environment_custom.rb'
before attempting to add a link in the shared links view. Without the
check, the links would always be available but would be dead ends on
the provider side if the admin hadn't provided a key and secret in the
omniauth provider config.