Add email validation back to User model
Validate the uniqueness of emails if one is provided during record
creation. Users authenticating with an oauth provider won't have an
email address so the field is allowed to be black. Those that do have
one need to have a unique address.

The index over user emails in the database allows for blank emails
across any number of user records. However, it loses the ability to
check for duplicate addresses where they do exist. Records then must
be checked in the model (but this is the only place they're validated).