Fix URl after idea creation
After an idea has been created, redirect to the idea's page with the
slug as the parameter for selecting the record to show. Otherwise, the
URL would have the record ID number as part of it after creation.
Showing records normally would still use the slugged URL, but
immediately after record creation it would still use the old record
IDs.