Reorganize location on main application assets
Move some assets (css, images) to an 'application' directory so that
admin resources can be easily excluded. This will prevent CSS
properties belonging to the admin interface being applied to anything
in the main application views. Main application images need to be moved
in a similar manner so that they can be properly referenced.

TODO:
  * Move and refrence main application javascripts in a similar manner.
  * Adjust main Admin asset referencing and inclusion to work in the
    same way the main application asset handling is done.